refactor: align cross-feature hook imports and type bulkAssignRole extra payload
- Clean up unreachable code in fetchIssuedLogs
- Refactor rule-form-dialog and extra-assignment-dialog to load hooks from public barrels
- Introduce and apply BulkAssignExtraPayload interface in bulkAssignRole and hook call sites
